Move route from `testing-routes.ts` to `manage-account.ts` Move logic from route into `manage-account.ts` controller. Add logic to update Cognito and our database.
