+++ express/src/routes/manage-account.ts
@@ -1,5 +1,5 @@
 import express, {Request, Response} from "express";
-import {listServices} from "../controllers/manage-account";
+import {changePassword, listServices, showChangePasswordForm} from "../controllers/manage-account";
 import {checkAuthorisation} from "../middleware/authoriser";
 import {showAddServiceForm, processAddServiceForm} from "../controllers/manage-account";
 import {serviceNameValidator} from "../middleware/serviceNameValidator";
@@ -40,6 +40,12 @@ router.get('/client-details/:serviceId', async (req, res) => {
     });
     req.session.updatedField = undefined;
 });
+
+router.get('/account/change-password', showChangePasswordForm);
+
+
+router.post('/change-password', changePassword);
+
 
 const DEFAULT_PUBLIC_KEY = 'MIIBojANBgkqhkiG9w0BAQEFAAOCAY8AMIIBigKCAYEAp2mLkQGo24Kz1rut0oZlviMkGomlQCH+iT1pFvegZFXq39NPjRWyatmXp/XIUPqCq9Kk8/+tq4Sgjw+EM5tATJ06j5r+35of58ATGVPniW//IhGizrv6/ebGcGEUJ0Y/ZmlCHYPV+lbewpttQ/IYKM1nr3k/Rl6qepbVYe+MpGubluQvdhgUYel9OzxiOvUk7XI0axPquiXzoEgmNNOai8+WhYTkBqE3/OucAv+XwXdnx4XHmKzMwTv93dYMpUmvTxWcSeEJ/4/SrbiK4PyHWVKU2BozfSUejVNhahAzZeyyDwhYJmhBaZi/3eOOlqGXj9UdkOXbl3vcwBH8wD30O9/4F5ERLKxzOaMnKZ+RpnygWF0qFhf+UeFMy+O06sdgiaFnXaSCsIy/SohspkKiLjNnhvrDNmPLMQbQKQlJdcp6zUzI7Gzys7luEmOxyMpA32lDBQcjL7KNwM15s4ytfrJ46XEPZUXESce2gj6NazcPPsrTa/Q2+oLS9GWupGh7AgMBAAE=';
